Natalie Bakes a Cake
'Natalie Bakes a Cake' is an algorithmic sequencing activity that teaches students on how to write an adventure story, and then code the story in Python. 

Learners can use the Algorithmic Board (A-Board) to tell the story of Natalie baking a delicious birthday cake for her friend's special day. The learner will then be able to interact with the story using PyCharm, and help Natalie pick out the ingredients and make some decisions that could take the story in an entirely new direction! 

Finally, Learners will be able to write their own stories, and code them in Python. With offline tools and online support, it's a piece of cake!