Open Source AI Research Paper
The Emerging Technology Final Report on Open Source Artificial Intelligence will dissect open source Artificial Intelligence (AI) and examine the advantages and disadvantages for companies to be so open by sharing their expertise and experiments. The development and adoption of AI-enabled technologies will transform how people live and work and interact. Technology companies that choose to develop open source AI want to scale and market share while seeking to build safe and beneficial tools that will maximize profits and enhance the quality of life for all. In exploring these topics, we researched the history of open source and the motivational drivers of both technology companies and the developers that contribute to the open source community. The authors of this report also developed and distributed a survey on open AI to gather insights on how experts believe AI will evolve and impact the industry, society, and the future workforce. The first half of the paper will cover the history of open source software, address why technology companies and software engineers choose to open source their code and make a case for open source AI. The second half of the paper will dive into the results of the Perceptions of Open AI survey and will explore major open-source Machine Learning software packages as well as open AI experiments.