Green Jobs and the Green Economy are Taking Shape
Deal or No Deal: How Green Jobs and the Green Economy are Taking Shape
As the 2020 electoral campaigns begin to build momentum, the call to address impacts of climate change are becoming part of the national conversation from how we can reduce carbon emissions to how we can stimulate economic growth through the creation of high-paying green jobs. Upgrading existing infrastructure and building a smart power grid, powered by artificial intelligence, will demand a specialized workforce, that requires reskilling, upskilling, and new types of credentialing to meet the current need. There is also short-term, high demand for professions in occupations that will support green infrastructure and industry. Lastly, there is a considerable amount of research and job creation in the renewable energy space that is highly specialized, but also highly lucrative.